refactor: folder-structure-page-admin-to-settings
